Increase the default FString limits
This enables ability to parse large FiBData blocks. Since these blocks can be essentially an arbitrary size may be it would be better to introduce default parameter for the function that would be used to ignore in-place limits. Since it is a default one, it will keep compatibility in tact and only requires change to internal to library inheriting members.