Fix Material Node Inputs for QualitySwitch/FeatureLevelSwitch
These serve to create more issues for interpreting the data, and since static arrays are parsed into Property[Index], this is no longer needed.

Issue: Ambiguous Input Index (if the last input is used in the original data, it's put as the first index)
Fix: Remove deserializer, it already works with the Property[Index] format, this just removes it in place of a ambiguous and improper implementation.