Update Psychonauts2.cs
I’ve renamed LinecodeTableIndex to TextsIndex, as it corresponds to a valid entry in the Texts array. The previous TextsIndex I've renamed to DialogueIndex, since its value can exceed the length of either the LinecodeTable or Texts arrays.

Each integer in a LinecodeTable entry represents an index, mapping the corresponding Linecode to a Data entry.